+++ packages/node/src/declarations.d.ts
@@ -1,1 +1,9 @@
 declare module 'lsmod';
+
+declare module 'https-proxy-agent' {
+  /** JSDoc */
+  // tslint:disable-next-line:no-unnecessary-class
+  export default class HttpsProxyAgent {
+    public constructor(proxy: string);
+  }
+}